requirements.txt
absl-py==0.7.0
astor==0.7.1
astroid==2.1.0
attr==0.3.1
attrs==18.2.0
audioread==2.1.6
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.3.2
ffmpeg==1.4
Flask==1.0.2
gast==0.2.2
google-api-core==1.8.1
google-auth==1.6.3
google-cloud-language==1.1.1
googleapis-common-protos==1.5.8
grpcio==1.18.0
h5py==2.9.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
librosa==0.6.3
llvmlite==0.27.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
mysqlclient==1.4.1
numba==0.42.1
numpy==1.16.1
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pylint==2.2.2
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
readline==6.2.4.1
requests==2.21.0
resampy==0.2.1
rsa==4.0
scikit-learn==0.20.2
scikits.samplerate==0.3.3
scipy==1.2.1
six==1.12.0
SQLAlchemy==1.2.17
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
typed-ast==1.1.1
urllib3==1.24.1
Wave==0.0.2
Werkzeug==0.14.1
wrapt==1.10.11
